Enable bootstrap customization of flux components
Allow users of the flux_sync data source to provide a list of patch
names to be applied to kustomization.yaml and be returned a
corresponding map of patch name keys to expected file location values.
This is to make tasks such as patching the kustomize-controller service
account to enable sops available as part of the bootstrap process.

Signed-off-by: Robert Farrimond <robert.farrimond@ovoenergy.com>